chore: move KeyValueInput to @atoms
